feat: create query controller
